Remove FetchAdvertisingId and simplify listener call in GetAdvertisingId
